put the config validator in its own file
